requirements.txt
asgiref==3.4.1
boto3==1.18.44
botocore==1.21.44
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
Collectfast==2.2.0
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.2.7
django-cors-headers==3.8.0
django-environ==0.7.0
django-heroku==0.3.1
django-redis-cache==1.6.5
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
django-storages==1.11.1
django-utils-six==2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
docutils==0.17.1
drf-nested-routers==0.93.3
gunicorn==20.1.0
idna==3.2
itypes==1.2.0
Jinja2==3.0.1
jmespath==0.10.0
MarkupSafe==2.0.1
mypy-extensions==0.4.3
openapi-codec==1.3.2
pathspec==0.9.0
platformdirs==2.3.0
psycopg2==2.9.1
PyJWT==2.1.0
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
regex==2021.8.28
requests==2.26.0
s3transfer==0.5.0
simplejson==3.17.5
six==1.16.0
sqlparse==0.4.2
tomli==1.2.1
typing-extensions==3.10.0.2
uritemplate==3.0.1
urllib3==1.26.6
whitenoise==5.3.0
